+++ CHANGELOG.md
@@ -5,6 +5,11 @@
     in between `$min` and `$max`.
   * `hypot()`: given *n* numbers, outputs the length of the *n*-dimensional
     vector that has components equal to each of the inputs.
+  * Exponential (all inputs must be unitless):
+    * `log($number)` or `log($number, $base)`. If no base is provided, `log()`
+      performs a natural log.
+    * `pow($base, $exponent)`
+    * `sqrt($number)`
 
 * Add the variables `$pi` and `$e` to the built-in "sass:math" module.